+++ lib/ui/laundry/LaundryRoomDetailPanel.dart
@@ -177,27 +177,33 @@ class _LaundryRoomDetailPanelState extends State<LaundryRoomDetailPanel> impleme
       Localization().getStringEx('widget.card.button.favorite.off.hint', '') :
       Localization().getStringEx('widget.card.button.favorite.on.hint', '');
 
-    String favoriteIcon = isFavorite? 'images/icon-star-blue.png' : 'images/icon-star-gray-frame-thin.png';
+    String favoriteIcon = isFavorite? 'images/icon-star-orange.png' : 'images/icon-star-white.png';
+    
     return Container(color: Colors.white, child:
-      Padding(padding: EdgeInsets.only(left: 24, right: 24, top: 11, bottom: 24), child:
+      Padding(padding: EdgeInsets.only(left: 24), child:
         Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
-          Row(mainAxisSize: MainAxisSize.max, mainAxisAlignment: MainAxisAlignment.spaceBetween, children: <Widget>[
-            Text(Localization().getStringEx('panel.laundry_detail.heading.laundry', 'Laundry'), style: TextStyle(fontFamily: Styles().fontFamilies?.bold, fontSize: 14, color: Styles().colors?.fillColorPrimary,),),
+          Row(children: <Widget>[
+            Expanded(child:
+              Padding(padding: EdgeInsets.symmetric(vertical: 12), child:
+                Text(widget.room.name ?? '', style: TextStyle( fontFamily: Styles().fontFamilies?.extraBold, fontSize: 22, color: Styles().colors?.fillColorPrimary,),),
+              ),
+            ),
             Visibility(visible: Auth2().canFavorite, child:
               GestureDetector(onTap: _onTapFavorite, child:
                 Semantics(label: favoriteLabel, hint: favoriteHint, button: true, excludeSemantics: true, child:
-                  Padding(padding: EdgeInsets.symmetric(vertical: 10), child:
+                  Padding(padding: EdgeInsets.only(left: 12, right: 16, top: 24, bottom: 25), child:
                     Image.asset(favoriteIcon, excludeFromSemantics: true)
                   ),
                 ),
               ),
             ),
           ],),
-          Padding(padding:EdgeInsets.only(top: 2, bottom: 14), child:
-            Text(widget.room.name ?? '', style: TextStyle( fontFamily: Styles().fontFamilies?.extraBold, fontSize: 24, color: Styles().colors?.fillColorPrimary,),),
+          Padding(padding: EdgeInsets.only(right: 24, bottom: 24), child:
+            Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
+              _buildLocationWidget(),
+              _buildReportIssueWidget()
+            ],)
           ),
-          _buildLocationWidget(),
-          _buildReportIssueWidget()
         ],),
       ),
     );