+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Handled "viewPoi" command in iOS MapView [#1699](https://github.com/rokwire/illinois-app/issues/1699).
+- Handled Laundry in Favorites and Browse panels [#1916](https://github.com/rokwire/illinois-app/issues/1916).
 ### Changed
 - Updated strings for phone / email sign up [#1931](https://github.com/rokwire/illinois-app/issues/1931).
+- Multiple updates for Laundry favorite / detail UI [#1916](https://github.com/rokwire/illinois-app/issues/1916).
 ### Fixed
 - Do not apply any logic whether to show laundry favorites in Saved panel [#1917](https://github.com/rokwire/illinois-app/issues/1917).